Fix for the nacl build failures on the WinBuild(dbg) builder.
BUG=none
TEST=it compiles
TBR=joi
Review URL: http://codereview.chromium.org/7324041

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@91981 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5276150edefd4d36110a3f420137853549c0c870

+++ content/common/notification_service.cc
@@ -119,7 +119,7 @@ NotificationService::~NotificationService() {
   lazy_tls_ptr.Pointer()->Set(NULL);
 
 #ifndef NDEBUG
-  for (size_t i = 0; i < observer_counts_.size(); i++) {
+  for (int i = 0; i < static_cast<int>(observer_counts_.size()); i++) {
     if (observer_counts_[i] > 0) {
       // This may not be completely fixable -- see
       // http://code.google.com/p/chromium/issues/detail?id=11010 .
